requirements.txt
pyopengl
numba
numpy
glfw